simple max advance
This is a minimalist change that has a pleasingly visible effect
on the behavior of the bot.  The point of this change is primarily
to serve as an example of a small change.  Secondarily, it is
kinda cool to see a simplistic but quite aggressive behavior.

1) create new package 'move_selection'
2) added function 'simple_max_advance' taking as input the array
   of legal moves and our own color (Color::WHITE or Color::BLACK).
3) 'simple_max_advance' selects the move that increases our rank
   (or decreases if BLACK) by the most.  When no advance is possible,
   select a random move.
4) Change 'send_challenge' to send a challenge to my other bot
   which has been running the base teacher model (random selection).

The following link shows this strategy in action running as BLACK.

https://lichess.org/iHC7cRqI