running C model
This model runs with a couple of bits commented out.
MC-3020 does not support assignment of arrays, so the assignment
of the legal_moves is commented out.
POSIX threading is turned on to be more friendly to OS and in anticipation
of running multiple threads in the future.
TIM is modified to allow Background to run periodically.